FE for initial decision and final decision
corrected naming conventions, applied the 'number steps' to initial decision and final decision pages. Adjusted the css for part of the heading.
