+++ pages/final-decision.tsx
@@ -3,10 +3,11 @@ import Heading from 'components/MashHeading/Heading';
 import NumberedSteps from 'components/NumberedSteps/NumberedSteps';
 import { Select } from 'components/Form';
 
-const ScreeningDecision = (): React.ReactElement => (
+const FinalDecision = (): React.ReactElement => (
   <>
-    <h1>Make screening decision</h1>
+    <h1>Make final decision</h1>
     <Heading clientname="Jan Smith" timeleft="3 hours" datetime="10:00 6 Jun" />
+
     <NumberedSteps
       nodes={[
         <>
@@ -18,16 +19,37 @@ const ScreeningDecision = (): React.ReactElement => (
             <a href="#" className="lbh-link lbh-link--no-visited-state">
               See Google document
             </a>
-          </p>{' '}
+          </p>
         </>,
         <>
-          {' '}
           <label htmlFor="screening-decision" className="lbh-heading-h3">
-            Select screening decision
+            Select final decision
           </label>
           <Select
-            id="screening-decision"
-            name="screening-decision"
+            id="initial-decision"
+            name="initial-decision"
+            ignoreValue
+            options={['option 1', 'option 2']}
+          />
+        </>,
+        <>
+          <label htmlFor="referral-category" className="lbh-heading-h3">
+            Select referral category
+          </label>
+          <Select
+            id="referral-category"
+            name="referral-category"
+            ignoreValue
+            options={['option 1', 'option 2']}
+          />
+        </>,
+        <>
+          <label htmlFor="referral-category" className="lbh-heading-h3">
+            Allocate out of MASH
+          </label>
+          <Select
+            id="referral-category"
+            name="referral-category"
             ignoreValue
             options={['option 1', 'option 2']}
           />
@@ -80,6 +102,7 @@ const ScreeningDecision = (): React.ReactElement => (
         </>,
       ]}
     />
+
     <div style={{ display: 'flex', alignItems: 'center' }}>
       <Button label="Submit" type="submit" />
       <p className="lbh-body">
@@ -95,11 +118,11 @@ const ScreeningDecision = (): React.ReactElement => (
           href="#"
           className={`lbh-link lbh-link--no-visited-state govuk-!-margin-left-3`}
         >
-          Return to initial decision
+          Return to contact
         </a>
       </p>
     </div>
   </>
 );
 
-export default ScreeningDecision;
+export default FinalDecision;
